backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.10.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
asyncpg==0.29.0 